+++ src/app/(createADao)/create/component/CreatePreviewModal/index.tsx
@@ -178,7 +178,7 @@ export default function CreatePreviewModal({ open, onClose, onConfirm }: ICreate
             descriptionList={[
               {
                 content: `
-                ${highCouncil?.highCouncilConfig.maxHighCouncilMemberCount} members and ${highCouncil?.highCouncilConfig.maxHighCouncilCandidateCount} candidates at most, rotated every ${highCouncil?.highCouncilConfig.electionPeriod} days. Require a staking of at least ${highCouncil?.highCouncilConfig.lockTokenForElection} ${metaData?.governanceToken} tokens.
+                ${highCouncil?.highCouncilConfig.maxHighCouncilMemberCount} members and ${highCouncil?.highCouncilConfig.maxHighCouncilCandidateCount} candidates at most, rotated every ${highCouncil?.highCouncilConfig.electionPeriod} days. Require a staking of at least ${highCouncil?.highCouncilConfig.stakingAmount} ${metaData?.governanceToken} tokens.
                 `,
               },
               {
