+++ src/config/index.ts
@@ -16,3 +16,4 @@ export const SECONDS_60 = 60000;
 export const ElectionContractName = 'Election';
 
 export const NetworkDaoHomePathName = '/network-dao';
+export const NetworkName = 'Network DAO';
